# MedicalDashboard/OpenAPI-documentation.yaml
openapi: 3.0.3
info:
  title: Medical Dashboard API
  description: |    
    ### Микросервисы:
    - **AuthService** - Аутентификация и управление пользователями
    - **DashboardAPI** - Управление пациентами, метриками и алертами
    - **NotificationService** - Отправка уведомлений
    
    ### Аутентификация
    Большинство эндпоинтов требуют JWT Bearer токен через `/identity/login`.
    
    ### SignalR
    Для real-time получения метрик используется SignalR Hub `/hubs/metrics` через Gateway.
    Gateway автоматически проксирует WebSocket соединения к DashboardAPI.
    Поддерживаются методы:
    - `SubscribeToPatient(patientId: Guid)` - подписка на метрики пациента
    - `SubscribeToPatientMetrics(patientId: Guid)` - подписка на метрики пациента (альтернативное имя)
    - `UnsubscribeFromPatient(patientId: Guid)` - отписка от метрик пациента
    
    ### Ограничения
    - Все запросы должны содержать заголовок `Content-Type: application/json`
    - JWT токены имеют ограниченное время жизни - 7 дней
  version: 1.0.0
servers:
  - url: http://localhost:7168
    description: Development Gateway (проксирует запросы к микросервисам)
security:
  - BearerAuth: []
paths:
...
  /patient-alerts:
    get:
      tags:
        - Alerts
      summary: Получить список алертов
      description: Получение списка алертов с фильтрацией и пагинацией
      operationId: getAlerts
      security:
        - BearerAuth: []
      parameters:
        - name: patientId
          in: query
          schema:
            type: string
            format: uuid
          description: Фильтр по ID пациента
        - name: isProcessed
          in: query
          schema:
            type: boolean
          description: Фильтр по статусу обработки
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Номер страницы
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: Размер страницы
      responses:
        '200':
          description: Список алертов
          headers:
            X-Total-Count:
              description: Общее количество алертов
              schema:
                type: integer
            X-Page:
              description: Текущая страница
              schema:
                type: integer
            X-PageSize:
              description: Размер страницы
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AlertDto'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Попытка доступа с недостаточными правами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /patient-alerts/{id}:
    get:
      tags:
        - Alerts
      summary: Получить алерт по ID
      description: Получение информации об алерте по его ID
      operationId: getAlert
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID алерта
      responses:
        '200':
          description: Информация об алерте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDto'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Попытка доступа с недостаточными правами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Алерт не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Alerts
      summary: Удалить алерт
      description: Удаление алерта по ID
      operationId: deleteAlert
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID алерта
      responses:
        '204':
          description: Алерт удален
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Попытка доступа с недостаточными правами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /patient-alerts/{id}/ack:
    post:
      tags:
        - Alerts
      summary: Подтвердить алерт
      description: Подтверждение обработки алерта
      operationId: acknowledgeAlert
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID алерта
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeAlertDto'
      responses:
        '200':
          description: Алерт подтвержден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertDto'
        '400':
          description: Неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неавторизованный доступ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Попытка доступа с недостаточными правами
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Алерт не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========================================
  # SIGNALR HUB ENDPOINTS
  # ========================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer токен для аутентификации
  schemas:
...
    AlertDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор алерта
          example: "123e4567-e89b-12d3-a456-426614174000"
        patientId:
          type: string
          format: uuid
          description: ID пациента
          example: "123e4567-e89b-12d3-a456-426614174001"
        alertType:
          type: string
          description: Тип алерта
          example: "critical"
        indicator:
          type: string
          description: Индикатор алерта
          example: "heart_rate"
        severity:
          type: string
          enum: [high, medium, low]
          description: Уровень серьезности
          example: "high"
        isProcessed:
          type: boolean
          description: Статус обработки
          example: false
        acknowledgedAt:
          type: string
          format: date-time
          nullable: true
          description: Время подтверждения
          example: "2024-01-15T10:30:00Z"
        acknowledgedBy:
          type: string
          format: uuid
          nullable: true
          description: ID пользователя, подтвердившего алерт
          example: "123e4567-e89b-12d3-a456-426614174002"
        createdAt:
          type: string
          format: date-time
          description: Время создания
          example: "2024-01-15T10:30:00Z"
    AcknowledgeAlertDto:
      type: object
      required:
        - acknowledgedBy
      properties:
        acknowledgedBy:
          type: string
          format: uuid
          description: ID пользователя, подтверждающего алерт
          example: "123e4567-e89b-12d3-a456-426614174002"

    # ========================================
    # NOTIFICATION SERVICE SCHEMAS
    # ========================================
...
    ErrorResponse:
      type: object
      properties:
        errorMessage:
          type: string
          description: Детальное сообщение об ошибке
          example: "Detailed error message"
...